Move column extensions to their own file
