+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/helpers/querybuilders/ColumnExtensions.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders
+
+object ColumnExtensions {
+  fun <T> Column<T>.cast(type: SqlType): CastColumn<T> = CastColumn(this, type)
+}
