+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/repository/deviceActivation/GetDeviceActivationByIdQueryBuilder.kt
@@ -1,13 +1,12 @@
 package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.repository.deviceActivation
 
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.ColumnExtensions.cast
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.JoinType
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.SqlType
-import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.cast
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.expressions.CurrentDate
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.model.athena.AthenaQuery
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.model.athena.DeviceActivation
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.model.athena.Person
-import java.time.ZonedDateTime
 
 class GetDeviceActivationByIdQueryBuilder(private val id: Long) {
   fun build(): AthenaQuery = DeviceActivation
@@ -24,7 +23,7 @@ class GetDeviceActivationByIdQueryBuilder(private val id: Long) {
     .where {
       DeviceActivation.deviceActivationId eq id
 
-      Person.groupedDate.cast<ZonedDateTime>(SqlType.Date) eq CurrentDate()
+      Person.groupedDate.cast(SqlType.Date) eq CurrentDate()
     }
     .prepare()
 }
