+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/repository/person/GetPersonByIdQueryBuilder.kt
@@ -1,11 +1,10 @@
 package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.repository.person
 
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.ColumnExtensions.cast
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.SqlType
-import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.cast
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.expressions.CurrentDate
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.model.athena.AthenaQuery
 import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.model.athena.Person
-import java.time.ZonedDateTime
 
 class GetPersonByIdQueryBuilder(private val id: String) {
   fun build(): AthenaQuery = Person
@@ -24,7 +23,7 @@ class GetPersonByIdQueryBuilder(private val id: String) {
     .where {
       Person.deviceWearerId eq id
 
-      Person.groupedDate.cast<ZonedDateTime>(SqlType.Date) eq CurrentDate()
+      Person.groupedDate.cast(SqlType.Date) eq CurrentDate()
     }
     .prepare()
 }
